@@ -6,11 +6,11 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node devserver.js",
-    "lint": "./node_modules/.bin/eslint app/**/*.js",
+    "lint": "./node_modules/.bin/eslint app/**/*.js app/**/**/*.js app/**/**/**/*.js webpack*js",
     "build-rules": "firebase-bolt firebase/security.bolt --output firebase/rules.json",
     "build-prod": "NODE_ENV=production webpack -p --config webpack.config.js",
     "build-dev": "NODE_ENV=development webpack -d --config webpack.config.js",
-    "format": "prettier --print-width 100 --bracket-spacing=false --write app/**/**/*.js app/**/*.js webpack*js"
+    "format": "prettier --print-width 100 --bracket-spacing=false --write app/**/**/*.js app/**/*.js app/**/**/**/*.js webpack*js"
   },
   "repository": {
     "type": "git",
@@ -23,10 +23,10 @@
   },
   "homepage": "https://github.com/zbuttram/treesradio-frontend#readme",
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/core": "^7.5.5",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/plugin-proposal-decorators": "^7.4.4",
-    "@babel/preset-env": "^7.4.5",
+    "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
     "babel-eslint": "^10.0.2",
     "babel-loader": "^8.0.6",
@@ -34,14 +34,14 @@
     "babel-polyfill": "^6.26.0",
     "clean-webpack-plugin": "^3.0.0",
     "ejs": "^2.6.2",
-    "eslint": "^6.0.0",
-    "eslint-plugin-react": "^7.14.0",
+    "eslint": "^6.1.0",
+    "eslint-plugin-react": "^7.14.3",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "^4.0.0",
+    "file-loader": "^4.1.0",
     "firebase-bolt": "^0.8.4",
     "font-awesome-webpack": "https://github.com/jarecsni/font-awesome-webpack",
     "git-rev-sync": "^1.12.0",
-    "graceful-fs": "^4.1.15",
+    "graceful-fs": "^4.2.0",
     "html-webpack-plugin": "^3.2.0",
     "html-webpack-template": "^6.2.0",
     "json-loader": "^0.5.7",
@@ -49,12 +49,12 @@
     "less-loader": "^5.0.0",
     "mobx-react-devtools": "^6.1.1",
     "node-sass": "^4.12.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "prettier": "^1.18.2",
     "sass-loader": "^7.1.0",
     "script-loader": "^0.7.2",
-    "webpack": "^4.35.0",
-    "webpack-cli": "^3.3.5",
+    "webpack": "^4.37.0",
+    "webpack-cli": "^3.3.6",
     "webpack-config": "^7.5.0",
     "webpack-dev-server": "^3.7.2",
     "webpack-version-file-plugin": "^0.4.0"
@@ -67,34 +67,34 @@
     "disposable-email": "^0.2.3",
     "emojione": "github:emojione/emojione",
     "favico.js": "^0.3.10",
-    "firebase": "^6.2.2",
+    "firebase": "^6.3.1",
     "font-awesome": "^4.7.0",
     "howler": "^2.1.2",
     "jquery": "^3.4.1",
     "localforage": "^1.7.3",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.15",
     "mitt": "^1.1.3",
-    "mobx": "^5.10.1",
+    "mobx": "^5.13.0",
     "mobx-react": "^6.1.1",
     "moment": "^2.24.0",
     "prop-types": "^15.7.2",
-    "rc-progress": "^2.4.1",
+    "rc-progress": "^2.5.1",
     "react": "^16.8.6",
     "react-bootstrap": "^1.0.0-beta.6",
     "react-dom": "^16.8.6",
     "react-emojione": "^5.0.1",
-    "react-markdown": "^4.0.9",
+    "react-markdown": "^4.1.0",
     "react-onclickoutside": "^6.8.0",
     "react-player": "^1.11.1",
     "react-slider": "^0.11.2",
-    "react-toastify": "^5.3.1",
-    "react-visibility-sensor": "^5.0.2",
+    "react-toastify": "^5.3.2",
+    "react-visibility-sensor": "^5.1.1",
     "url-parse": "^1.4.7"
   },
   "babel": {
     "presets": [
-      "@babel/env",
-      "@babel/react"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ],
     "plugins": [
       [
