@@ -1,24 +1,19 @@
 import React from "react";
 import Modal from "../../components/utility/Modal";
 import {observer} from "mobx-react";
-import {observable, computed} from "mobx";
 
 @observer
 export default class Karan extends React.Component {
-  componentWillMount() {
+  constructor(props, context) {
+    super(props, context);
     this.state = {
-      modalIsOpen: true
+      modalIsOpen: props.isVisible,
+      hideCallback: props.hideCallback
     };
-  }
-  @computed
-  get toggleModal() {
-    this.setState({
-      modalIsOpen: !this.state.modalIsOpen
-    });
   }
   render() {
     return (
-      <Modal show={this.state.modalIsOpen} hideModal={() => this.toggleModal} title="In Memoriam">
+      <Modal show={this.state.modalIsOpen} title="In Memoriam">
         <div className="row">
           <div className="col-md-3">
             <img src="//i.imgur.com/f9cyJz3.jpg" width="100%" />
@@ -31,7 +26,7 @@ export default class Karan extends React.Component {
               Saeft, or Karan as some of us came to know him, was a wonderful human being and a
               great friend to many of us here at TR. His presence was one of love, respect, and a
               lightheartedness that was unmatched. His love of music and the laughs he shared with
-              us is something that we're all going to miss dearly.
+              us is something that we&apos;re all going to miss dearly.
             </p>
             {/* <p>If you or someone you know is suffering from depression or suicidal thoughts, please seek help immediately. 
                             Self medication has its limits. All problems can be managed, and nobody has to suffer alone.</p>
