@@ -1,16 +1,11 @@
-import React from 'react';
-import rank from 'libs/rank';
-import getUsername from 'libs/username';
-import classNames from 'classnames';
+import React from "react";
+import rank from "libs/rank";
+import getUsername from "libs/username";
+import classNames from "classnames";
 
 const noop = () => {};
 
 export default class UserName extends React.Component {
-	componentWillUnmount() {
-		this.setState = (state,callback)=>{
-		return;
-		};
-	}
   constructor(props) {
     super(props);
     this.state = {};
@@ -18,32 +13,33 @@ export default class UserName extends React.Component {
     this.getUsername();
   }
 
-  getRank = async() => {
+  getRank = async () => {
     const title = await rank(this.props.uid);
     this.setState({title});
-  }
+  };
 
-  getUsername = async() => {
+  getUsername = async () => {
     const username = await getUsername(this.props.uid);
     this.setState({username});
-  }
+  };
 
   render() {
-    let userClass = 'username-user';
-    if ( this.state.username == 'BlazeBot' ) userClass = 'username-bot';
+    let userClass = "username-user";
+    if (this.state.username == "BlazeBot") userClass = "username-bot";
 
     if (this.state.title) {
-      userClass = `username-${this
-        .state
-        .title
-        .split(' ')
-        .join('')
+      userClass = `username-${this.state.title
+        .split(" ")
+        .join("")
         .toLowerCase()}`;
     }
 
     const usernameClasses = classNames(this.props.className, userClass);
 
-    return <span onClick={this.props.onClick || noop} className={usernameClasses}>{this.state.username}</span>;
+    return (
+      <span onClick={this.props.onClick || noop} className={usernameClasses}>
+        {this.state.username}
+      </span>
+    );
   }
-
 }
